requirements.txt
python-telegram-bot==11.1.0
dialogflow==0.6.0
vk-api==11.4.0
python-dotenv==0.10.2